requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asyncio==4.0.0
attrs==25.4.0
certifi==2025.11.12
EbookLib==0.20
edge-tts==7.2.7
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
idna==3.11
lxml==6.0.2
multidict==6.7.0
propcache==0.4.1
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
yarl==1.22.0
